requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
playwright==1.21.0
tqdm==4.65.0
sentence-transformers==2.2.2
faiss-cpu==1.7.3
requests==2.28.1
pydantic==1.10.6
